% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot_Cell_Proportion.R
\name{Barplot_Cell_Proportion}
\alias{Barplot_Cell_Proportion}
\title{Bar plot of the cell proportions in each identity}
\usage{
Barplot_Cell_Proportion(
  seurat_object,
  group.by = NULL,
  split.by = NULL,
  idents = NULL,
  group.idents = NULL,
  split.idents = NULL,
  colors = NULL,
  order.prop = NULL,
  order.group = NULL,
  order.split = NULL,
  order.colors = TRUE,
  alpha = 1,
  show.cellsum.label = TRUE,
  cellsum.label.size = 3,
  axis.text.size = 9,
  x.axis.angle = 60,
  x.axis.hjust = 1,
  y.axis.title.size = 11,
  legend.text.size = 9,
  legend.side = "bottom",
  show.legend = TRUE,
  split.plot.title.size = 24,
  prop.percent = TRUE,
  nrow = 1,
  unique.group.plot = TRUE,
  unique.split.plot = FALSE,
  output.data = FALSE
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{group.by}{Character. The name of a metadata (for example, 'orig.ident', 'seurat_clusters', etc) to group the identities of the active.ident metadata by. Returns a stacked bar plot.}

\item{split.by}{Character. The name of a metadata to split the identities of the active.ident metadata by. Returns separate \code{ggplot2} objects.}

\item{idents}{Character. The names of one or several identities in the active.ident metadata to select. If \code{NULL}, all identities are used.}

\item{group.idents}{Character. The names of one or several \code{group.by} identities to select. If \code{NULL}, all identities are used. Ignored if \code{group.by} = \code{NULL}.}

\item{split.idents}{Character. The names of one or several \code{split.by} identities to select. If \code{NULL}, all identities are used. Ignored if \code{split.by} = \code{NULL}.}

\item{colors}{Character. The color names for each identity of the active.ident metadata or in \code{idents}. If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{order.prop}{Character or Numeric. Either 'reverse', or the identities (as names or as numeric values corresponding to the indices) of the active.ident metadata or in \code{idents} to order the cells.}

\item{order.group}{Character. Either 'reverse', or the names of the \code{group.by} identities or in \code{group.idents} to order the cells. Ignored if \code{group.by} = \code{NULL}.}

\item{order.split}{Character. Either 'reverse', or the names of the \code{split.by} identities or in \code{split.idents} to order the cells. Ignored if \code{split.by} = \code{NULL}.}

\item{order.colors}{Logical. If \code{TRUE}, the \code{colors} will automatically be ordered according to \code{order.prop}. Ignored if \code{order.prop} = \code{NULL}.}

\item{alpha}{Numeric. The transparency of the bar colors. A value between 0 and 1.}

\item{show.cellsum.label}{Logical. If \code{TRUE}, the cell sum will be shown at the top of each bar.}

\item{cellsum.label.size}{Numeric. The font size of the cell sum label. Ignored if \code{show.cellsum.label} = \code{FALSE}.}

\item{axis.text.size}{Numeric. The font size of the identity names and cell percent or count values.}

\item{x.axis.angle}{Numeric. The rotation angle of the identity names.}

\item{x.axis.hjust}{Numeric. The horizontal justification of the identity names.}

\item{y.axis.title.size}{Numeric. The font size of the y axis title.}

\item{legend.text.size}{Numeric. The font size of the legend text. Ignored if \code{show.legend} = \code{FALSE}.}

\item{legend.side}{Character. The side where the legend will be displayed, either 'left', 'right', 'top' or 'bottom'. Ignored if \code{show.legend} = \code{FALSE}.}

\item{show.legend}{Logical. If \code{TRUE}, shows the legend.}

\item{split.plot.title.size}{Numeric. The font size of the split plot titles. Ignored if \code{split.by} = \code{NULL}.}

\item{prop.percent}{Logical. If \code{TRUE}, the cell proportions will be shown as a percentage of total cells for each identity. Ignored if \code{group.by} = \code{NULL}.}

\item{nrow}{Numeric. The number of rows in the \code{patchwork} object. Ignored if \code{group.by} = \code{NULL} and \code{split.by} = \code{NULL}.}

\item{unique.group.plot}{Logical. If \code{TRUE}, the stacked bar plots will be gathered in a single \pkg{ggplot2} object. Ignored if \code{group.by} = \code{NULL}.}

\item{unique.split.plot}{Logical. If \code{TRUE}, the \pkg{ggplot2} objects will be gathered in a single \code{patchwork}. Ignored if \code{split.by} = \code{NULL}.}

\item{output.data}{Logical. If \code{TRUE}, the function will return a \code{data.frame} object or a \code{list} of \code{data.frame} objects of the cell proportions, instead of displaying anything.}
}
\value{
A \pkg{ggplot2} object, a \code{list} of \pkg{ggplot2} objects, a \code{patchwork} object containing \pkg{ggplot2} objects or a \code{list} of \code{patchwork} objects containing \pkg{ggplot2} objects.
}
\description{
This function generates a bar plot of the cell proportions from a \pkg{Seurat} object.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")

# Example 1: default parameters
Barplot_Cell_Proportion(pbmc3k)
}
