% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimSmoothers.R
\name{dimSmoothers}
\alias{dimSmoothers}
\title{Scatterplot of the dimensionality reduction and fitted GAM smoothers for each cell}
\usage{
dimSmoothers(
  sds,
  models = NULL,
  sds.conditions = NULL,
  genes = NULL,
  lineages = NULL,
  clusters = NULL,
  conditions = NULL,
  knots = NULL,
  show.knots.labels = FALSE,
  rescale = FALSE,
  rescale.range = c(0, 3),
  pseudotime.type = "relative",
  points.size = 1,
  lineages.width = 1,
  lineages.arrow = arrow(),
  knots.size = lineages.width * 3,
  cells.colors = if (is.character(genes)) "turbo" else if (any(clusters == "pseudotime"))
    "viridis" else NULL,
  na.color = "lightgrey",
  show.na = TRUE,
  lineages.colors = "grey50",
  knots.colors = lineages.colors,
  knots.labels.text.colors = "white",
  knots.labels.fill.colors = lineages.colors,
  axis.text.size = 9,
  axis.title.size = 11,
  plot.title.size = 16,
  lineages.legend.names = NULL,
  legend.text.size = 9,
  raster = FALSE
)
}
\arguments{
\item{sds}{A \pkg{SingleCellExperiment} object containing the pseudotime values of one or several lineages, computed using \code{\link[slingshot]{slingshot}} (usually, the input object to \code{\link[tradeSeq]{fitGAM}}).}

\item{models}{A \pkg{SingleCellExperiment} object containing the fitted GAM smoothers, computed using \code{\link[tradeSeq]{fitGAM}}, with or without \code{conditions} provided.}

\item{sds.conditions}{A list of \pkg{SingleCellExperiment} objects obtained using \code{\link[condiments]{slingshot_conditions}}. The conditions used to compute the pseudotime values must be the same as those used to fit the GAM smoothers in the \code{models} object, if provided.}

\item{genes}{Character. The names of one or several genes to plot the fitted GAM smoothers for each cell from. If the \code{models} object is not provided, the function will plot the log-transformed counts.}

\item{lineages}{Numeric. The indices of the lineages (for example, c(1, 5), 2:4 etc) to plot the pseudotime values from. Set to \code{NA} to remove all lineages.}

\item{clusters}{Character or Factor. Either the name of a metadata present in the \code{sds} object (for example, 'annotations', 'seurat_clusters', etc) to color the cells, or the identities, as character or factor, of length equal to the number of cells in the \code{sds} object. Set to 'pseudotime' to color the cells by their pseudotime values.}

\item{conditions}{Character. The names of one or several \code{conditions} identities to select. If \code{NULL}, all identities are used, and each unique condition will be plotted, for each of the \code{lineages} provided (for example, if you have four lineages and three conditions, twelve curves will be plotted). Ignored if the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}. Please note that if the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} with \code{conditions}, it is not possible to plot global lineages (without \code{conditions}), you would need to compute a new \code{models} object using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}. This is due to a limitation in how \code{\link[tradeSeq]{predictSmooth}} returns average fitted GAM smoothers (if the \code{models} object was computed with \code{conditions}, the function will always mean the fitted GAM smoothers for each lineage and each condition independently).}

\item{knots}{Numeric. The indices of the knots to display on the \code{lineages}. Set to \code{NA} to remove all knots. Ignored if the \code{models} object is not provided.}

\item{show.knots.labels}{Logical. If \code{TRUE}, the knot numbers will be displayed using \code{\link[ggrepel]{geom_label_repel}}. Ignored if the \code{models} object is not provided.}

\item{rescale}{Logical. If \code{TRUE}, fitted GAM smoothers will be adjusted using \code{\link[scales]{rescale}} between the first numerical value of \code{rescale.range} (lowest value) and the second numerical value (highest value). This is different than \code{\link[base]{scale}} as this doesn't compare values to any mean or standard deviation and is therefore not a Z-score, it only refits each value (independently for each gene) in order to visualize all \code{genes} in the same dimension regardless of their differences in fitted GAM smoothers.}

\item{rescale.range}{Numeric. The minimum and maximum values to resize the fitted GAM smoothers and internally passed to \code{\link[scales]{rescale}}. These values are arbitrary and will not change the visualization, only the values in the legend. Ignored if \code{rescale} = \code{FALSE}.}

\item{pseudotime.type}{Character. Determines the \code{cells.colors} range scale when \code{clusters} = 'pseudotime' and not all \code{lineages} are displayed: either 'absolute', where the \code{cells.colors} range is based on the highest pseudotime value among the displayed \code{lineages}, or 'relative', where the \code{cells.colors} range is based on the highest pseudotime value of all the \code{lineages}, including the non-displayed ones.}

\item{points.size}{Numeric. The size of the cells.}

\item{lineages.width}{Numeric. The width of the lineage curves.}

\item{lineages.arrow}{An \code{\link[grid]{arrow}} object added at the end of the lineage curves. Use the function's parameters to modify its appearance. Set to \code{NULL} to remove the arrow.}

\item{knots.size}{Numeric. The size of the knots on the lineage curves. Ignored if the \code{models} object is not provided.}

\item{cells.colors}{Character. If \code{genes} is provided, either two color names to use for the cells, corresponding to the lowest and highest values in the fitted GAM smoothers and internally passed to \code{\link[ggplot2]{scale_color_gradient}}, or the name of a palette and internally passed to \code{\link[ggplot2]{scale_color_viridis_c}} (such as 'turbo', 'inferno', 'magma', 'viridis' etc, check the function for all palettes available). If \code{clusters} is provided, the color names for each identity in \code{clusters}. If \code{NULL}, uses \pkg{Seurat}'s default colors. If \code{clusters} = 'pseudotime', the name of a palette and internally passed to \code{\link[ggplot2]{scale_color_viridis_c}}. If neither \code{genes} nor \code{clusters} are provided, the name of a color to use for all cells. If \code{NULL}, defaults to 'lightgrey'.}

\item{na.color}{Character. The name of a color to use for cells not participating in the \code{lineages} (for example, if you have four lineages and you only plot lineages 1 and 3, the cells participating in lineages 2 and 4 will be colored with \code{na.color}). Set to 'transparent' to not display these cells.}

\item{show.na}{Logical. If \code{FALSE}, ignores \code{na.color} and colors all cells with \code{cells.colors}, even those not participating in the \code{lineages}.}

\item{lineages.colors}{Character. either a single color name to use for all lineages, or the color names for each of the \code{lineages} displayed, or for the number of \code{lineages} times the number of \code{conditions} if the \code{models} object was computed with \code{conditions}.}

\item{knots.colors}{Character. either a single color name to use for all knots, or a vector of color names of length equal to the number of \code{lineages}. Ignored if the \code{models} object is not provided.}

\item{knots.labels.text.colors}{Character. either a single color name to use for all knot labels, or a vector of color names of length equal to the number of \code{lineages}. Ignored if the \code{models} object is not provided or if \code{show.knots.labels} = \code{FALSE}.}

\item{knots.labels.fill.colors}{Character. either a single color name to use for all knot label backgrounds, or a vector of color names of length equal to the number of \code{lineages}. Ignored if the \code{models} object is not provided or if \code{show.knots.labels} = \code{FALSE}.}

\item{axis.text.size}{Numeric. The font size of the dimensionality reduction values.}

\item{axis.title.size}{Numeric. The font size of the axis title.}

\item{plot.title.size}{Numeric. The font size of the plot title.}

\item{lineages.legend.names}{Character. You may provide custom names for the lineages displayed in the legend. Ignored if \code{lineages.colors} is a single color.}

\item{legend.text.size}{Numeric. The font size of the legend text.}

\item{raster}{Logical. If \code{TRUE}, the cells will be rasterized using \code{\link[ggrastr]{rasterize}} to reduce the size of the output file. The points may appear slightly blurry.}
}
\value{
A \pkg{ggplot2} object, or a \code{patchwork} object containing \pkg{ggplot2} objects.
}
\description{
This function generates a scatterplot of the log-transformed counts, and displays on top the average fitted GAM smoother curves, computed using \code{\link[tradeSeq]{fitGAM}}, along pseudotime, and is a reworked version of \code{\link[tradeSeq]{plotSmoothers}}, it allows for the plotting of multiple genes simultaneously, as well as selected lineages and conditions. Additionally, the scatterplot is rasterized to reduce the size of the output file and branching points may be displayed on each curve. Finally, the log-transformed counts and average fitted GAM smoother curves are rescaled to the same range between genes to allow for a better comparison.
}
