% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotPlot_Heatmap.R
\name{DotPlot_Heatmap}
\alias{DotPlot_Heatmap}
\title{Dot plot or heatmap of average gene expression in each identity}
\usage{
DotPlot_Heatmap(
  seurat_object,
  assay = "RNA",
  layer = "data",
  data.are.log = TRUE,
  features,
  group.by = NULL,
  split.by = NULL,
  annotation.vars = NULL,
  separate.split = TRUE,
  annotation.type = "absolute",
  annotation.plot = c("donut", "violin"),
  annotation.split = TRUE,
  idents = NULL,
  split.idents = NULL,
  scale = TRUE,
  rescale = FALSE,
  rescale.range = c(0, 3),
  rotate.axis = FALSE,
  dotplot = TRUE,
  dots.type = "square root",
  dots.size = 5,
  show.noexpr.dots = FALSE,
  col.min = ifelse(isTRUE(scale), -2, 0),
  col.max = ifelse(isTRUE(scale), 2, "q100"),
  data.colors = if (isTRUE(scale)) c("#35A5FF", "white", "red") else "Viridis",
  palette.reverse = FALSE,
  na.color = "grey40",
  background.color = "white",
  background.alpha = 0.7,
  idents.colors = NULL,
  idents.colors.outer.border = TRUE,
  idents.colors.inner.border = FALSE,
  show.idents.names.colors = FALSE,
  show.idents.oppo.colors = TRUE,
  split.colors = NULL,
  split.colors.outer.border = TRUE,
  split.colors.inner.border = FALSE,
  show.split.names.colors = FALSE,
  show.split.oppo.colors = TRUE,
  annotation.colors = NULL,
  annotation.colors.outer.border = TRUE,
  annotation.colors.inner.border = FALSE,
  order.idents = NULL,
  order.split = NULL,
  order.annotation = NULL,
  order.colors = TRUE,
  kmeans.repeats = 100,
  cluster.idents = TRUE,
  idents.kmeans = 1,
  idents.kmeans.numbers.size = 11,
  cluster.features = TRUE,
  features.kmeans = 1,
  features.kmeans.numbers.size = 11,
  idents.gap = 1,
  features.gap = 1,
  idents.names.size = 9,
  features.names.size = 9,
  features.names.style = "italic",
  row.names.side = "left",
  row.names.width = 15,
  column.names.angle = 45,
  column.names.side = "bottom",
  column.names.height = 15,
  inner.border = TRUE,
  outer.border = TRUE,
  data.legend.name = ifelse(isTRUE(scale), "Z-Score", "Average Expression"),
  data.legend.side = "bottom",
  data.legend.direction = "horizontal",
  data.legend.position = "topcenter",
  data.legend.width = 5,
  show.data.legend = TRUE,
  idents.legend.name = "Clusters",
  show.idents.legend = TRUE,
  split.legend.name = split.by,
  show.split.legend = TRUE,
  annotation.legend.name = annotation.vars,
  show.annotation.legend = TRUE,
  legend.title.size = 10,
  legend.text.size = 10,
  legend.gap = 10,
  heatmap.width = 10,
  heatmap.height = 10,
  output.data = FALSE,
  ...
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{assay}{Character. The name of an assay containing the \code{layer} with the expression matrix. If the \code{seurat_object} contains multiple 'RNA' assays, you may specify which one to use (for example, 'RNA2' if you have created a second 'RNA' assay you named 'RNA2'. See \href{https://satijalab.org/seurat/articles/seurat5_essential_commands.html#create-seurat-or-assay-objects}{Seurat v5 vignettes} for more information). You may also use another assay, such as 'SCT', to pull feature expression from.}

\item{layer}{Character. The name of a layer (formerly known as slot) which stores the expression matrix. It is recommended to use 'data'.}

\item{data.are.log}{Logical. If \code{TRUE}, and if \code{layer} = 'data', the function assumes data are log-transformed and cell expression values will be exponentiated (using \code{\link[base]{expm1}}) so that averaging is done in non-log space (as per \code{\link[Seurat]{DotPlot}} or \code{\link[Seurat]{AverageExpression}}'s default behavior), average expression values are then log-transformed back (using \code{\link[base]{log1p}}). If \code{layer} = 'scale.data' or 'counts', cell expression values are not exponentiated prior to averaging.}

\item{features}{Character. The names of one or several features to plot the average expression from.}

\item{group.by}{Character. The name of a metadata (for example, 'orig.ident', 'seurat_clusters', etc) to use instead of the active.ident metadata. If \code{NULL}, the current active.ident metadata will be used.}

\item{split.by}{Character. The name of a metadata (for example, 'orig.ident', 'seurat_clusters', etc) to split the identities of the active.ident metadata by.}

\item{annotation.vars}{Character. The names of one or several metadata (categorical variables) or features (continuous variables) to display as annotation variables.}

\item{separate.split}{Logical. If \code{TRUE}, \code{cluster.features} will be set to \code{FALSE} and the different \code{split.by} identities will split the \code{features} instead of the identities of the active.ident metadata. Ignored if \code{split.by} = \code{NULL}.}

\item{annotation.type}{Character. Either 'absolute', which will calculate the proportion of cells within the whole object (for example, with some visualizations like 'bar', identities with low amount of cells will appear small), or 'relative', which will calculate the proportion of cells independently in each identity of the active.ident metadata, all identities will be displayed with the same relative size, which might distort the perception of the true proportion of cells, so use with caution. Only applies to categorical variables. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{annotation.plot}{Character. The type of plot to use for annotation variables. First for the categorical variables, either 'donut', 'pie', 'treemap', 'treemap_fixed', 'bar' or 'bar_fixed'. Then for the continuous variables, either 'violin', 'violin_median', 'violin_quartiles', 'violin_boxplot', 'boxplot' or 'density'. You must choose any one of each (for example, c('donut', 'violin'), or c('treemap', 'boxplot') etc). You may also provide multiple choices of each if the length matches the number of annotation variables (for example, if you have 3 annotation variables, you may set \code{annotation.plot} = c('treemap', 'boxplot', 'donut'), or 5 annotation variables \code{annotation.plot} = c('pie', 'violin_boxplot', 'violin_boxplot', 'density', 'bar') etc). You need to know beforehand which variables are categorical and which are continuous to set the corresponding choice. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{annotation.split}{Logical. If \code{TRUE}, the annotation variables will be split by the \code{split.by} identities. You may also provide multiple choices if the length matches the number of annotation variables (if you have 3 annotation variables, you may set \code{annotation.split} = c(TRUE, FALSE, TRUE) etc). Ignored if \code{annotation.vars} = \code{NULL} or if \code{split.by} = \code{NULL}.}

\item{idents}{Character. The names of one or several identities in the active.ident metadata to select. If \code{NULL}, all identities are used.}

\item{split.idents}{Character. The names of one or several \code{split.by} identities to select. If \code{NULL}, all identities are used. Ignored if \code{split.by} = \code{NULL}.}

\item{scale}{Logical. If \code{TRUE}, average expression values will be scaled using \code{\link[base]{scale}} and default parameters. The resulting values will be Z-scores (mean subtracted values divided by standard deviation) and not positive average expression values anymore, which is why there will be positive and negative values displayed, depending on if the average expression in a particular identity is below or above the mean average expression from all identities (which is calculated independently for each feature). Caution should be exercised when interpreting results with low number of identities (typically below five), as small differences in average expression might lead to exacerbated differences when scaled.}

\item{rescale}{Logical. If \code{TRUE}, average expression values will be adjusted using \code{\link[scales]{rescale}} between the first numerical value of \code{rescale.range} (lowest expression) and the second numerical value (highest expression). This is different than \code{\link[base]{scale}} as this doesn't compare values to any mean and standard deviation and is therefore not a Z-score, it only refits each average expression value (independently for each feature) in order to visualize all \code{features} in the same dimension regardless of their differences in levels of expression. Caution should be exercised when interpreting results with low number of identities (typically below five), as small differences in average expression might lead to exacerbated differences when rescaled. Ignored if \code{scale} = \code{TRUE}.}

\item{rescale.range}{Numeric. The minimum and maximum values to resize the average expression values and internally passed to \code{\link[scales]{rescale}}. These values are arbitrary and will not change the visualization, only the values in the legend, you need to adjust \code{col.min} and \code{col.max} to influence the color scale. Ignored if \code{rescale} = \code{FALSE} or \code{scale} = \code{TRUE}.}

\item{rotate.axis}{Logical. If \code{TRUE}, flips the axis, so that \code{features} are displayed as rows and identities as columns.}

\item{dotplot}{Logical. If \code{TRUE}, the function will display a dot plot, with dots size proportional to the percentage of cells in each identity expressing a feature. If \code{FALSE}, the function will instead display a heatmap.}

\item{dots.type}{Character. Determines the dot size differences between 0 and 100\% expression. Either 'square root' (lower difference) or 'radius' (higher difference). Ignored if \code{dotplot} = \code{FALSE}.}

\item{dots.size}{Numeric. The size of the dots. Decreasing this parameter helps when displaying a large number of \code{features}. Ignored if \code{dotplot} = \code{FALSE}.}

\item{show.noexpr.dots}{Logical. If \code{TRUE}, the function will display a small dot for \code{features} with 0\% expression, instead of nothing. Ignored if \code{dotplot} = \code{FALSE}.}

\item{col.min}{Character or Numeric. The minimum value for the \code{breaks} internally passed to \code{\link[colorRamp2]{colorRamp2}}. If character, must be a quantile in the form 'qX' where X is a number between 0 and 100. A value of 'q5' or 'q10' is useful to reduce the effect of outlier values (i.e. a very low value that significantly alters the color scale range of all other values).}

\item{col.max}{Character or Numeric. The maximum value for the \code{breaks} internally passed to \code{\link[colorRamp2]{colorRamp2}}. If character, must be a quantile in the form 'qX' where X is a number between 0 and 100. A value of 'q95' or 'q90' is useful to reduce the effect of outlier values (i.e. a very high value that significantly alters the color scale range of all other values).}

\item{data.colors}{Character or Function. Either three color names, corresponding to the lowest, zero (or middle if \code{scale} = \code{FALSE}), and highest values in the expression matrix and internally passed to \code{\link[colorRamp2]{colorRamp2}}, or two color names, corresponding to the lowest and highest values, or the name of a palette and internally passed to \code{hcl_palette} in \code{\link[colorRamp2]{colorRamp2}} (such as 'Inferno', 'Berlin', 'Viridis' etc, check \code{\link[grDevices]{hcl.pals}} for all palettes available), or a custom \code{\link[colorRamp2]{colorRamp2}} function.}

\item{palette.reverse}{Logical. If \code{TRUE} and if \code{data.colors} is a palette (such as 'Viridis'), the function will reverse its colors.}

\item{na.color}{Character. The color name for missing values (\code{NA}).}

\item{background.color}{Character. Either a color name or 'data.colors', 'idents.colors' or 'split.colors' for the background behind the dots. Ignored if \code{dotplot} = \code{FALSE}.}

\item{background.alpha}{Numeric. The alpha for the background color. Ignored if \code{dotplot} = \code{FALSE}.}

\item{idents.colors}{Character. The color names for each identity of the active.ident metadata or in \code{idents}. If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{idents.colors.outer.border}{Logical. If \code{TRUE}, the function will display a border around each identity slice.}

\item{idents.colors.inner.border}{Logical. If \code{TRUE}, the function will display a border around each identity row or column.}

\item{show.idents.names.colors}{Logical. If \code{TRUE}, the function will display the colors specified in \code{idents.colors} next to identity names.}

\item{show.idents.oppo.colors}{Logical. If \code{TRUE}, the function will display the colors specified in \code{idents.colors} on the opposite side of identity names.}

\item{split.colors}{Character. The color names for each \code{split.by} identity or in \code{split.idents}. If \code{NULL}, uses a custom set of colors from \code{\link[grDevices]{colors}}. Ignored if \code{split.by} = \code{NULL}.}

\item{split.colors.outer.border}{Logical. If \code{TRUE}, the function will display a border around each \code{split.by} identity slice. Ignored if \code{split.by} = \code{NULL}.}

\item{split.colors.inner.border}{Logical. If \code{TRUE}, the function will display a border around each \code{split.by} identity row or column. Ignored if \code{split.by} = \code{NULL}.}

\item{show.split.names.colors}{Logical. If \code{TRUE}, the function will display the colors specified in \code{split.colors} next to identity names. Ignored if \code{split.by} = \code{NULL}.}

\item{show.split.oppo.colors}{Logical. If \code{TRUE}, the function will display the colors specified in \code{split.colors} on the opposite side of identity names. Ignored if \code{split.by} = \code{NULL}.}

\item{annotation.colors}{Character or List. The color names for each annotation variable identity. You may also provide a named list if you want to set the colors of multiple annotation variables (for example, if \code{annotation.vars} = c('treatment', 'response'), you may provide \code{annotation.colors} = list('treatment' = c('blue', 'red', 'gold'), 'response' = c('magenta', 'green', 'brown', 'lightblue'))). For categorical variables, the length must match the number of identities in the annotation variable. For continuous variables, the length must match the number of active.ident identities if \code{annotation.split} = FALSE, or the number of \code{split.by} identities if \code{annotation.split} = TRUE. If \code{NULL}, uses a custom set of colors from \code{\link[grDevices]{colors}}. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{annotation.colors.outer.border}{Logical. If \code{TRUE}, the function will display a border around each annotation variable slice. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{annotation.colors.inner.border}{Logical. If \code{TRUE}, the function will display a border around each annotation variable row or column. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{order.idents}{Character or Numeric. Either 'reverse', or the identities (as names or as numerical values corresponding to the indices) of the active.ident metadata or in \code{idents} to order the active.ident identities. If \code{cluster.idents} = \code{TRUE} or Function, only the legend names will be ordered.}

\item{order.split}{Character or Numeric. Either 'reverse', or the \code{split.by} identities (as names or as numerical values corresponding to the indices) or in \code{split.idents} to order the \code{split.by} identities. If \code{cluster.idents} = \code{TRUE} or Function, only the legend names will be ordered. Ignored if \code{split.by} = \code{NULL}.}

\item{order.annotation}{Character or Numeric. Either 'reverse', or the annotation variable identities (as names or as numerical values corresponding to the indices) to order the annotation variable identities. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{order.colors}{Logical. If \code{TRUE}, the \code{data.colors} and \code{split.colors} will automatically be ordered according to \code{order.idents} and \code{order.split}. Ignored if \code{order.idents} and \code{order.split} are \code{NULL}.}

\item{kmeans.repeats}{Numeric. The number of runs to get a consensus K-means clustering. Ignored if \code{idents.kmeans} and \code{features.kmeans} are equal to 1.}

\item{cluster.idents}{Logical or Function. If \code{TRUE}, the function will cluster the identities. You may also pass an \code{hclust} or \code{dendrogram} object which contains clustering.}

\item{idents.kmeans}{Numeric. The number of slices to use for identity K-means clustering.}

\item{idents.kmeans.numbers.size}{Numeric. The font size of the identity K-means slice numbers. Set to 0 to remove them.}

\item{cluster.features}{Logical or Function. If \code{TRUE}, the function will cluster the \code{features}. You may also pass an \code{hclust} or \code{dendrogram} object which contains clustering.}

\item{features.kmeans}{Numeric. The number of slices to use for feature K-means clustering.}

\item{features.kmeans.numbers.size}{Numeric. The font size of the feature K-means slice numbers. Set to 0 to remove them.}

\item{idents.gap}{Numeric. The gap between the identity slices. Ignored if \code{idents.kmeans} = 1.}

\item{features.gap}{Numeric. The gap between the feature slices. Ignored if \code{features.kmeans} = 1.}

\item{idents.names.size}{Numeric. The font size of the identity names. Set to 0 to remove them.}

\item{features.names.size}{Numeric. The font size of the feature names. Set to 0 to remove them.}

\item{features.names.style}{Character. The font face of the feature names. The \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7494048/}{Gene nomenclature} used by almost all scientific journals require that feature names are italicized, therefore the parameter is by default set to 'italic'. Use 'plain' to revert back to regular font face.}

\item{row.names.side}{Character. The side where the row names will be displayed, either 'left' or 'right'. The dendrogram will be displayed on the opposite side.}

\item{row.names.width}{Numeric. The width of the row names. Increase this parameter if your row names are truncated.}

\item{column.names.angle}{Numeric. The angle of rotation of the column names.}

\item{column.names.side}{Character. The side where the column names will be displayed, either 'top' or 'bottom'. The dendrogram will be displayed on the opposite side.}

\item{column.names.height}{Numeric. The height of the column names. Increase this parameter if your column names are truncated.}

\item{inner.border}{Logical. If \code{TRUE}, the function will display a black outline around each dot if \code{dotplot} = \code{TRUE}, or a black border around each cell of the heatmap if \code{dotplot} = \code{FALSE}.}

\item{outer.border}{Logical. If \code{TRUE}, the function will display an outer border around the plot or around each slice if \code{idents.kmeans} and/or \code{features.kmeans} are higher than 1.}

\item{data.legend.name}{Character. The name of the data legend.}

\item{data.legend.side}{Character. The side where the data legend will be displayed, either 'left', 'right', 'top' or 'bottom'.}

\item{data.legend.direction}{Character. The direction of the data legend, either 'horizontal' or 'vertical'.}

\item{data.legend.position}{Character. The centering of the data legend name, there are many options, default option from \code{\link[ComplexHeatmap]{Heatmap}} is 'topleft'.}

\item{data.legend.width}{Numeric. How long the data legend will be, only affects the data legend if \code{data.legend.direction} = 'horizontal'.}

\item{show.data.legend}{Logical. If \code{TRUE}, the function will display a legend for the average expression data.}

\item{idents.legend.name}{Character. The name of the active.ident metadata legend. Ignored if \code{show.idents.names.colors} and \code{show.idents.oppo.colors} are \code{FALSE}.}

\item{show.idents.legend}{Logical. If \code{TRUE}, the function will display a legend for the active.ident metadata identities or \code{idents}. Ignored if \code{show.idents.names.colors} and \code{show.idents.oppo.colors} are \code{FALSE}.}

\item{split.legend.name}{Character. The name of the \code{split.by} legend. Ignored if \code{split.by} = \code{NULL}. Ignored if \code{show.split.names.colors} and \code{show.split.oppo.colors} are \code{FALSE}.}

\item{show.split.legend}{Logical. If \code{TRUE}, the function will display a legend for \code{split.by} identities or \code{split.idents}. Ignored if \code{show.split.names.colors} and \code{show.split.oppo.colors} are \code{FALSE}.}

\item{annotation.legend.name}{Character. The name of the annotation variable legend, the length must match the number of annotation variables. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{show.annotation.legend}{Logical. If \code{TRUE}, the function will display a legend for annotation variable identities. Ignored if \code{annotation.vars} = \code{NULL}.}

\item{legend.title.size}{Numeric. The font size of all legend titles.}

\item{legend.text.size}{Numeric. The font size of all legend texts.}

\item{legend.gap}{Numeric. The gap between the legends and the dot plot or heatmap. This parameter sets the value in the global options of \code{\link[ComplexHeatmap]{ht_opt}}, so it will affect all \code{\link[ComplexHeatmap]{Heatmap}} objects in the same R session. Use \pkg{ComplexHeatmap}::\code{\link[ComplexHeatmap]{ht_opt}}(RESET = \code{TRUE}) to restore default parameters.}

\item{heatmap.width}{Numeric. The width of each column.}

\item{heatmap.height}{Numeric. The height of each row.}

\item{output.data}{Logical. If \code{TRUE}, the function will return a \code{list} containing a \code{matrix} object of the average expression data, scaled or not, and another \code{matrix} object containing the percentage of cells expressing each feature, instead of displaying anything.}

\item{...}{Additional arguments to be passed to \code{\link[ComplexHeatmap]{Heatmap}}, such as \code{show_parent_dend_line}, \code{clustering_method_rows}, etc, accepts any parameter that wasn't already internally passed to \code{\link[ComplexHeatmap]{Heatmap}} (for example, \code{outer.border} sets the \code{border} parameter of \code{\link[ComplexHeatmap]{Heatmap}}, so you will get an error if you try to pass the \code{border} parameter in \code{\link[RightOmicsTools]{DotPlot_Heatmap}}).}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap}} object, either as a dot plot, or a heatmap, or a \code{list} containing a \code{matrix} object of the average expression data, scaled or not, and another \code{matrix} object containing the percentage of cells expressing each feature.
}
\description{
This function generates a dot plot or a heatmap to visualize the average expression of features in each identity of the active.ident metadata of a \pkg{Seurat} object. Credits to \pkg{Seurat}'s dev team for the original \code{\link[Seurat]{DotPlot}} from which data processing of this function is derived from and to \href{https://divingintogeneticsandgenomics.com/post/clustered-dotplot-for-single-cell-rnaseq/}{Ming Tang} for the initial idea to use \pkg{ComplexHeatmap} to draw a dot plot and the \code{layer_fun} function that draws the dots. Various new parameters were added to offer more flexibility and customization.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")
pbmc3k.markers <- c("CCR7", "CD14", "CD40LG",
                    "CD79A", "CD8A", "CDKN1C",
                    "GNLY", "CLEC10A", "PPBP")

# Example 1: default parameters
DotPlot_Heatmap(pbmc3k,
                features = pbmc3k.markers)
}
