% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_VlnPlot.R
\name{Grid_VlnPlot}
\alias{Grid_VlnPlot}
\title{Square stacked violin plot of gene expression in each identity}
\usage{
Grid_VlnPlot(
  seurat_object,
  assay = "RNA",
  layer = "data",
  features,
  idents = NULL,
  scale = TRUE,
  rotate.axis = FALSE,
  colors = NULL,
  order.idents = NULL,
  order.colors = TRUE,
  idents.text.size = 9,
  show.idents = FALSE,
  features.text.size = 9,
  legend.text.size = 7,
  legend.side = "bottom",
  show.legend = TRUE,
  ncol = "square"
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{assay}{Character. The name of an assay containing the \code{layer} with the expression matrix. If the \code{seurat_object} contains multiple 'RNA' assays, you may specify which one to use (for example, 'RNA2' if you have created a second 'RNA' assay you named 'RNA2'. See \href{https://satijalab.org/seurat/articles/seurat5_essential_commands.html#create-seurat-or-assay-objects}{Seurat v5 vignettes} for more information). You may also use another assay, such as 'SCT', to pull feature expression from.}

\item{layer}{Character. The name of a layer (formerly known as slot) which stores the expression matrix. It is recommended to use 'data'.}

\item{features}{Character. The names of one or several features to plot the cell expression from.}

\item{idents}{Character. The names of one or several identities in the active.ident metadata to select. If \code{NULL}, all identities are used.}

\item{scale}{Logical. If \code{TRUE}, scales the violins to have the same max height between features.}

\item{rotate.axis}{Logical. If \code{TRUE}, flips the axis, displaying violins vertically instead of horizontally.}

\item{colors}{Character. The color names for each identity of the active.ident metadata or in \code{idents}. If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{order.idents}{Character or Numeric. Either 'reverse', or the identities (as names or as numeric values corresponding to the indices) of the active.ident metadata or in \code{idents} to order the cells.}

\item{order.colors}{Logical. If \code{TRUE}, the \code{colors} will automatically be ordered according to \code{order.prop}. Ignored if \code{order.prop} = \code{NULL}.}

\item{idents.text.size}{Numeric. The font size of the identity names. Ignored if \code{show.idents} = \code{FALSE}.}

\item{show.idents}{Logical. If \code{TRUE}, shows the identity names on the plot.}

\item{features.text.size}{Numeric. The font size of the feature names.}

\item{legend.text.size}{Numeric. The font size of the legend text. Ignored if \code{show.legend} = \code{FALSE}.}

\item{legend.side}{Character. The side where the legend will be displayed, either 'left', 'right', 'top' or 'bottom'. Ignored if \code{show.legend} = \code{FALSE}.}

\item{show.legend}{Logical. If \code{TRUE}, shows the legend.}

\item{ncol}{Numeric. The number of columns to use. If 'square', will display features in a square grid or as close as possible depending on the number of features.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
This function is a stacked violin plot optimized to display the average expression of features in a \pkg{Seurat} object in a grid fashion (square) instead of a single column like other stacked violin functions available in other packages.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")
pbmc3k.markers <- c("CCR7", "CD14", "CD40LG",
                    "CD79A", "CD8A", "CDKN1C",
                    "GNLY", "CLEC10A", "PPBP")

# Example 1: default parameters
Grid_VlnPlot(pbmc3k,
             features = pbmc3k.markers)
}
