% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncRNA_human}
\alias{ncRNA_human}
\title{ncRNA_human}
\format{
\subsection{ncRNA_human}{

A character vector with 14574 elements.
}
}
\source{
\url{https://www.genenames.org/data/genegroup/#!/group/475}
}
\usage{
ncRNA_human
}
\description{
A character vector containing all identified non coding RNA symbols in the human genome as of may 10th 2024.
}
\keyword{datasets}
