% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncRNA_mouse}
\alias{ncRNA_mouse}
\title{Mouse non-coding genes}
\format{
\subsection{ncRNA_mouse}{

A character vector with 55720 elements.
}
}
\source{
\url{https://ftp.ebi.ac.uk/pub/databases/gencode/Gencode_mouse/release_M38/gencode.vM38.annotation.gtf.gz}
}
\usage{
ncRNA_mouse
}
\description{
A character vector containing all known non-coding RNA genes in the mouse genome as of december 12th 2025.
}
\keyword{datasets}
