% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudogenes_mouse}
\alias{pseudogenes_mouse}
\title{Mouse pseudogenes}
\format{
\subsection{pseudogenes_mouse}{

A character vector with 3663 elements.
}
}
\source{
\url{https://rna.sysu.edu.cn/dreamBase2/scrna.php?SClade=mammal&SOrganism=mm10&SDataId=0&SProteinID=0}
}
\usage{
pseudogenes_mouse
}
\description{
A character vector containing all identified pseudogenes in the mouse genome as of may 10th 2024.
}
\keyword{datasets}
