% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reference_Matrix_Builder.R
\name{Reference_Matrix_Builder}
\alias{Reference_Matrix_Builder}
\title{Builder of CIBERSORTx Reference Matrix}
\usage{
Reference_Matrix_Builder(
  seurat_object,
  assay = "RNA",
  layer = "counts",
  ident.1 = NULL,
  ident.2 = NULL,
  double.ident = TRUE,
  double.ident.sep = "_",
  reverse.double.ident = FALSE,
  subset.ident.1 = NULL,
  subset.ident.1.invert = FALSE,
  subset.ident.2 = NULL,
  subset.ident.2.invert = FALSE,
  downsample.object.first = NULL,
  downsample.object.last = NULL,
  downsample.cluster = NULL,
  automatic.downsample = FALSE,
  check.size = FALSE,
  max.matrix.size = 500,
  cell.barcodes = FALSE,
  file.name = "Reference_Matrix",
  file.format = "txt",
  file.sep = "\\t",
  write.path = NULL,
  write = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{assay}{Character. The name of an assay containing the \code{layer} with the expression matrix. If the \code{seurat_object} contains multiple 'RNA' assays, you may specify which one to use (for example 'RNA2' if you have created a second 'RNA' assay you named 'RNA2'. See \href{https://satijalab.org/seurat/articles/seurat5_essential_commands.html#create-seurat-or-assay-objects}{Seurat v5 vignettes} for more information). You may also use another assay, such as 'SCT', if you want to extract the expression matrix for projects other than CIBERSORTx.}

\item{layer}{Character. The name of a layer (formerly known as slot) which stores the expression matrix. Must be 'counts' layer for CIBERSORTx. You may also specify another layer such as 'data' (log-normalized counts) or any other present in the specified \code{assay} if you want to extract the expression matrix for projects other than CIBERSORTx. If you have split layers the function will always join them before extracting the expression matrix. The raw counts matrix will be normalized to Counts Per Million (CPM) for CIBERSORTx.}

\item{ident.1}{Character. The name of a metadata (for example, 'orig.ident', 'seurat_clusters', etc) from which to extract identities as column names in the Reference Matrix. If \code{NULL}, uses the active.ident metadata.}

\item{ident.2}{Character. The name of a metadata from which to split \code{ident.1} identities by (for example if \code{ident.1} is 'seurat_clusters' and \code{ident.2} is 'orig.ident', identity 0 will be separated into several 0_identities, each corresponding to an orig.ident: 0_sample1, 0_sample2 etc).}

\item{double.ident}{Logical. If \code{TRUE}, column names of the Reference Matrix will be set with dual identities (ident.1_ident.2). If \code{FALSE}, column names will be set as \code{ident.1} identities. Ignored if \code{ident.2} = \code{NULL}.}

\item{double.ident.sep}{Character. The separator to use between the identity names of \code{ident.1} and \code{ident.2}. Ignored if \code{ident.2} = \code{NULL} or \code{double.ident} = \code{FALSE}.}

\item{reverse.double.ident}{Logical. If \code{TRUE}, the identity names of \code{ident.1} and \code{ident.2} will be reversed (ident.2_ident.1). Ignored if \code{ident.2} = \code{NULL} or \code{double.ident} = \code{FALSE}.}

\item{subset.ident.1}{Character. The names of one or several \code{ident.1} identities to select. If \code{NULL}, all identities are used.}

\item{subset.ident.1.invert}{Logical. If \code{TRUE}, inverts the selection from \code{subset.ident.1} (for example, if \code{subset.ident.1} = c('0','1','2') and \code{subset.ident.1.invert} = \code{TRUE}, all identities except 0, 1 and 2 will be kept). Ignored if \code{subset.ident.1} = \code{NULL}.}

\item{subset.ident.2}{Character. The names of one or several \code{ident.2} identities to select. If \code{NULL}, all identities are used. Ignored if \code{ident.2} = \code{NULL}.}

\item{subset.ident.2.invert}{Logical. If \code{TRUE}, inverts the selection from \code{subset.ident.2} (for example, if \code{subset.ident.2} = c('patient1','patient4') and \code{subset.ident.2.invert} = \code{TRUE}, all identities except patient 1 and patient 4 will be kept). Ignored if \code{subset.ident.2} = \code{NULL} or \code{ident.2} = \code{NULL}.}

\item{downsample.object.first}{Numeric. The number of cells to downsample from the entire \code{seurat_object} (not from each identity) before subsetting with \code{subset.ident.1} and/or \code{subset.ident.2}. Please note that the relative proportion of each identity will be kept (for example if an identity is 10\% of total cells, it will remain 10\% after downsampling). If \code{NULL}, all cells are used.}

\item{downsample.object.last}{Numeric. The number of cells to downsample from the entire \code{seurat_object} after subsetting with \code{subset.ident.1} and/or \code{subset.ident.2}. Please note that the relative proportion of each identity will be kept (for example if an identity is 10\% of total cells, it will remain 10\% after downsampling). If \code{NULL}, all cells are used.}

\item{downsample.cluster}{Numeric. The number of cells to downsample from each \code{ident.1} identity. Will be performed after \code{downsample.object.last}. If \code{NULL}, all cells are used.}

\item{automatic.downsample}{Logical. If \code{TRUE}, automatically downsamples the \code{seurat_object} (respecting the relative proportion of each \code{ident.1} identity, it is therefore similar to \code{downsample.object.first} and \code{downsample.object.last}) so that the Reference Matrix file written to disk would be just under the \code{max.matrix.size} limit (empirically estimated). Performed last. Ignored if \code{check.size} = \code{TRUE} or \code{write} = \code{FALSE}. Please \href{https://github.com/Alexis-Varin/RightOmicsTools/issues}{report an issue} if you see a significant difference between the file size written to disk and \code{max.matrix.size} (for example, \code{max.matrix.size} is set to 200 MB but the file is 400 MB or 100 MB).}

\item{check.size}{Logical. If \code{TRUE}, prints the estimated size of the Reference Matrix file that would be written to disk and the number of cells to downsample if need be.}

\item{max.matrix.size}{Numeric. The maximum size of the Reference Matrix file written to disk in MB. Will stop the function if the Reference Matrix file written to disk is estimated to be over this limit, or if \code{automatic.downsample} = \code{TRUE}, will downsample the \code{seurat_object} instead so that the Reference Matrix written to disk is under the size limit. Ignored if \code{write} = \code{FALSE}.}

\item{cell.barcodes}{Logical. Must be \code{FALSE} for CIBERSORTx. If \code{TRUE}, keeps the cell barcodes and does not rename with cell identities, if you want to extract the expression matrix for projects other than CIBERSORTx.}

\item{file.name}{Character. The name of the Reference Matrix file written to disk. Must not contain any space for CIBERSORTx, the function will automatically replace any space with an underscore. Ignored if \code{check.size} = \code{TRUE} or \code{write} = \code{FALSE}.}

\item{file.format}{Character. The format of the Reference Matrix file written to disk. Must be 'txt' or 'tsv' for CIBERSORTx, but you may also specify 'csv' for example, if you want to extract the expression matrix for projects other than CIBERSORTx. Accepts any format the \code{\link[data.table]{fwrite}} function would accept. Ignored if \code{check.size} = \code{TRUE} or \code{write} = \code{FALSE}.}

\item{file.sep}{Character. The separator to use in the Reference Matrix file written to disk. Must be tabulation for CIBERSORTx, but you may also specify a comma for example, if you want to extract the expression matrix for projects other than CIBERSORTx. Accepts any separator the \code{\link[data.table]{fwrite}} function would accept. Ignored if \code{check.size} = \code{TRUE} or \code{write} = \code{FALSE}.}

\item{write.path}{Character. The path to write the Reference Matrix into. If \code{NULL}, uses current working directory. Ignored if \code{check.size} = \code{TRUE} or \code{write} = \code{FALSE}.}

\item{write}{Logical. If \code{TRUE}, writes to disk the Reference Matrix file. Ignored if \code{check.size} = \code{TRUE}.}

\item{verbose}{Logical. If \code{FALSE}, does not print progress messages and output, but warnings and errors will still be printed.}
}
\value{
A \code{data.table} object containing the normalized counts to CPM from the \code{seurat_object} or any other specified \code{assay} and \code{layer}, with cell identities or barcodes as column names and feature names as first column. If \code{write} = \code{TRUE}, the \code{data.table} object is also written to disk. If \code{check.size} = \code{TRUE}, will instead print the estimated size of the Reference Matrix file that would be written to disk and the number of cells to downsample if need be.
}
\description{
This function builds a Reference Matrix for \href{https://cibersortx.stanford.edu/index.php}{CIBERSORTx} from a \pkg{Seurat} object.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")

# Example 1:
refmat <- Reference_Matrix_Builder(pbmc3k,
                                   ident.1 = "seurat_annotations",
                                   ident.2 = "orig.ident",
                                   downsample.object.first = 1500,
                                   subset.ident.1 = c("Naive CD4 T",
                                                      "Memory CD4 T",
                                                      "CD8 T",
                                                      "NK"),
                                   subset.ident.2 = "Donor_2",
                                   subset.ident.2.invert = TRUE,
                                   write = FALSE)
refmat[1:5, 1:5]

# Example 2:
Reference_Matrix_Builder(pbmc3k,
                         check.size = TRUE,
                         max.matrix.size = 40)

}
