% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Right_Data.R
\name{Right_Data}
\alias{Right_Data}
\title{Data loader for vignettes}
\usage{
Right_Data(dataset = NULL, list.datasets = FALSE)
}
\arguments{
\item{dataset}{Character. The name of the dataset to load.}

\item{list.datasets}{Logical. If TRUE, prints the list of available datasets.}
}
\value{
An object of various classes, depending on the dataset asked, or a message containing the list of available datasets.
}
\description{
This function downloads and prepares data used for vignettes, such as \code{Seurat} or \code{SingleCellExperiment} objects. The main advantage of using a data loader is that nothing is stored in the package, therefore it remains lightweight even if you use large objects in your tutorials. Anyone interested in adding a dataset to this function for use in their own package's vignettes (provided the raw data can be downloaded from a reputable source, \href{https://ftp.ncbi.nlm.nih.gov/geo/datasets/}{NCBI's GEO FTP site} for example) can open an \href{https://github.com/Alexis-Varin/RightSeuratTools/issues}{issue on GitHub}.
}
\examples{
Right_Data(list.datasets = TRUE)

pbmc3k = Right_Data("pbmc3k")
pbmc3k
}
