% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Right_DietSeurat.R
\name{Right_DietSeurat}
\alias{Right_DietSeurat}
\title{Right_DietSeurat}
\usage{
Right_DietSeurat(
  seurat_object,
  idents = NULL,
  cells = NULL,
  features = NULL,
  dimreducs = NULL,
  graphs = NULL,
  variable.features = FALSE,
  misc = TRUE,
  split.counts.layer = NULL,
  data.layer = FALSE,
  scale.layer = FALSE,
  SCTAssay = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{idents}{Character. A vector with one or several identity names in the meta.data slot to keep in the diet object. If NULL, all identities will be kept.}

\item{cells}{Character. A vector with one or several cell barcodes to keep in the diet object. If NULL, all cells will be kept.}

\item{features}{Character. A vector with one or several feature names to keep in the diet object. If NULL, all features will be kept.}

\item{dimreducs}{Character. A vector with one or several reduction names to keep in the diet object. If NULL, all DimReducs will be removed. Note that if you subset features and remove all from which PCs were calculated, the PCA will not be kept even if it is specified in this parameter as the feature.loadings slot will be empty, this is intended behavior.}

\item{graphs}{Character. A vector with one or several graph names to keep in the diet object. If NULL, all Graphs will be removed.}

\item{variable.features}{Logical. If TRUE, the variable features slot will be kept in the diet object. Note that if you keep only a subset of features, the variable features will also be subset, and if you remove all features from which variable features were determined, the variable features will not be kept even if this parameter is set to TRUE, this is intended behavior.}

\item{misc}{Logical. If TRUE, the misc slot will be kept in the diet object.}

\item{split.counts.layer}{Character. The name of the identity to split the counts layers if you need to in your downstream analysis. If NULL, the diet object will have a single counts layer, even if the original object had split counts layers.}

\item{data.layer}{Logical. If TRUE, the data layer of the RNA assay will be kept in the diet object if it is present. As with the counts layer, if there are split data layers, they will be joined into a single data layer unless the split.counts.layer parameter is specified.}

\item{scale.layer}{Logical. If TRUE, the scale.data layer of the RNA assay will be kept in the diet object if it is present. As with the counts layer, if there are split scale.data layers, they will be joined into a single scale.data layer unless the split.counts.layer parameter is specified.}

\item{SCTAssay}{Logical. If TRUE, the SCT assay will be kept in the diet object if it is present.}
}
\value{
A Seurat object, hopefully smaller, with class Assay5 RNA assay and specified layers and slots.
}
\description{
As of Seurat v5 release, Seurat's DietSeurat() function does not remove data and scale.data layers, resulting in objects with little to no reduction in size. This function was created with the purpose to restore DietSeurat()'s proper behavior until the function is fixed by Seurat's dev team, as well as to offer a few new options such as the ability to subset cells or keep variable features.
}
\examples{
library(Seurat)

# Original object from SeuratObject:

pbmc_small

# Example 1:

Right_DietSeurat(pbmc_small,
                idents = c("letter.idents","groups"),
                features = c("CD3D", "NOSIP", "SAFB2",
                 "CD2", "IL7R", "PIK3IP1", "MAL"),
                graphs = "RNA_snn",
                split.counts.layer = "letter.idents")

# Example 2:

pbmc_cells = WhichCells(pbmc_small, idents = c("0","1"))

Right_DietSeurat(pbmc_small,
                cells = pbmc_cells,
                variable.features = TRUE,
                dimreducs = "pca",
                data.layer = TRUE)

}
