% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeSeqTests.R
\name{tradeSeqTests}
\alias{tradeSeqTests}
\title{Perform \pkg{tradeSeq} tests in parallel}
\usage{
tradeSeqTests(
  models,
  tests = c("assoc", "pattern", "diffend", "early", "startvsend", "condition"),
  global = TRUE,
  pairwise = TRUE,
  lineages = TRUE,
  l2fc = 0,
  eigen.thresh = 0.01,
  n.points = 2 * nknots(models),
  knots = NULL,
  pseudotime.values = NULL,
  parallelized = FALSE,
  BPPARAM = NULL,
  tidy = TRUE,
  raw = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{models}{A \pkg{SingleCellExperiment} object containing the fitted GAM smoothers, computed using \code{\link[tradeSeq]{fitGAM}}, with or without conditions provided.}

\item{tests}{Character. The names of one or several tests to perform. You may provide partial names, such as 'assoc' for \code{\link[tradeSeq]{associationTest}}, as long as a single match is found. The available tests are \code{\link[tradeSeq]{associationTest}}, \code{\link[tradeSeq]{patternTest}}, \code{\link[tradeSeq]{diffEndTest}}, \code{\link[tradeSeq]{earlyDETest}}, \code{\link[tradeSeq]{startVsEndTest}} and \code{\link[tradeSeq]{conditionTest}}. \code{\link[tradeSeq]{conditionTest}} is ignored if \code{models} was computed using \code{\link[tradeSeq]{fitGAM}} without conditions.}

\item{global}{Logical. (from \code{\link[tradeSeq]{earlyDETest}}/\code{\link[tradeSeq]{startVsEndTest}} documentation) If \code{TRUE}, test for all pairwise comparisons/lineages simultaneously.}

\item{pairwise}{Logical. (from \code{\link[tradeSeq]{earlyDETest}} documentation) If \code{TRUE}, test for all pairwise comparisons independently. Ignored in tests that do not have \code{pairwise}.}

\item{lineages}{Logical. (from \code{\link[tradeSeq]{startVsEndTest}} documentation) If \code{TRUE}, test for all lineages independently. Ignored in tests that do not have \code{lineages}.}

\item{l2fc}{Numeric. (from \code{\link[tradeSeq]{earlyDETest}} documentation) The log2 fold change threshold to test against. Note, that this will affect both the global test and the pairwise comparisons.}

\item{eigen.thresh}{Numeric. (from \code{\link[tradeSeq]{earlyDETest}} documentation) Eigenvalue threshold for inverting the variance-covariance matrix of the coefficients to use for calculating the Wald test statistics. Lower values are more lenient to adding more information but also decrease computational stability. This argument should in general not be changed by the user but is provided for back-compatability. Set to 1e-8 to reproduce results of older version of 'tradeSeq'. Ignored in tests that do not have \code{eigen.thresh}.}

\item{n.points}{Numeric. (from \code{\link[tradeSeq]{earlyDETest}} documentation) The number of points to be compared between lineages. Defaults to twice the number of knots. Ignored in tests that do not have \code{n.points}.}

\item{knots}{Numeric or List. (from \code{\link[tradeSeq]{earlyDETest}} documentation) A vector of length 2 specifying the knots at the start and end of the region of interest. You may also provide a \code{list} of multiple elements (for example, list(c(2,4), c(3,4))) to repeat the test for multiple sets of knots. Ignored in tests that do not have \code{knots}.}

\item{pseudotime.values}{Numeric or List. (from \code{\link[tradeSeq]{startVsEndTest}} documentation) A vector of length 2, specifying two pseudotime values to be compared against each other, for every lineage of the trajectory. @details Note that this test assumes that all lineages start at a pseudotime value of zero, which is the starting point against which the end point is compared. You may also provide a \code{list} of multiple elements (for example, list(c(8,12), c(6,14))) to repeat the test for multiple sets of pseudotime values. Ignored in tests that do not have \code{pseudotime.values}.}

\item{parallelized}{Logical. If \code{TRUE}, the \code{tests} will be parallelized using \pkg{BiocParallel}. Please note that parallelization is complex and depends on your operating system (Windows users might not see a gain or might even experience a slowdown).}

\item{BPPARAM}{A \code{\link[BiocParallel]{BiocParallelParam}} object to be used for parallelization. If \code{NULL} and \code{parallelized} = \code{TRUE}, the function will use a \code{\link[BiocParallel]{SerialParam}} object configured to use a single worker (core) and is therefore not parallelized, in order to prevent accidental use of large computation resources. Ignored if \code{parallelized} = \code{FALSE}.}

\item{tidy}{Logical. If \code{TRUE}, a \code{list} is returned, with \code{data.frame} objects corresponding to each test divided into each global, pairwise and/or lineage comparison results. An adjusted p-value (False Discovery Rate) is also calculated and each \code{data.frame} object is ordered based on decreasing Waldstat.}

\item{raw}{Logical. If \code{TRUE}, a \code{list} is returned, with \code{data.frame} objects corresponding to each test results.}

\item{verbose}{Logical. If \code{FALSE}, does not print progress messages and output, but warnings and errors will still be printed.}
}
\value{
A \pkg{list}, with \code{data.frame} objects corresponding to each test results and/or \code{data.frame} objects corresponding to each test and each global, pairwise and/or lineage comparison results.
}
\description{
This function is a wrapper around \pkg{tradeSeq} test functions, such as \code{\link[tradeSeq]{associationTest}} or \code{\link[tradeSeq]{conditionTest}}, and allows multiple tests to be performed simultaneously, with all necessary parameters specified internally. It also organizes the output of each test into separate \code{list} elements, making it easy to access results. Finally, it calculates an adjusted p-value and orders based on decreasing Waldstat.
}
