% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot_Cell_Proportion.R
\name{Barplot_Cell_Proportion}
\alias{Barplot_Cell_Proportion}
\title{Barplot_Cell_Proportion}
\usage{
Barplot_Cell_Proportion(
  seurat_object,
  group.by = NULL,
  split.by = NULL,
  colors = NULL,
  order.proportion = NULL,
  order.group = NULL,
  order.split = NULL,
  order.colors = TRUE,
  show.cellsum.label = TRUE,
  cellsum.label.size = 3,
  axis.text.size = 9,
  x.axis.angle = 60,
  x.axis.hjust = 1,
  y.axis.title.size = 11,
  legend.text.size = 9,
  legend.position = "bottom",
  split.plot.title.size = 24,
  legend = TRUE,
  percent = TRUE,
  nrow = 1,
  unique.group.plot = TRUE,
  unique.split.plot = FALSE
)
}
\arguments{
\item{seurat_object}{A Seurat object.}

\item{group.by}{Character. The name of an identity to group the active ident by into a stacked barplot.}

\item{split.by}{Character. The name of an identity to split the active ident by into separate ggplot objects.}

\item{colors}{Character. A vector of colors to be used. Seurat's default colors will be used if NULL.}

\item{order.proportion}{Character. A vector specifying either "reverse" or the levels of the active ident to order the cells.}

\item{order.group}{Character. A vector specifying either "reverse" or the levels of the group.by identity to order the cells. Ignored if group.by is NULL.}

\item{order.split}{Character. A vector specifying either "reverse" or the levels of the split.by identity to order the cells. Ignored if split.by is NULL.}

\item{order.colors}{Logical. If TRUE, the colors will automatically be ordered according to order.proportion. Ignored if order.proportion is NULL.}

\item{show.cellsum.label}{Logical. If TRUE, the sum of cells of each identity will be shown at the top of each bar.}

\item{cellsum.label.size}{Numeric. The size of the cell sum label. Ignored if show.cellsum.label is FALSE.}

\item{axis.text.size}{Numeric. The size of the x axis identities and y axis cell percent or numbers.}

\item{x.axis.angle}{Numeric. The angle of the x axis identities.}

\item{x.axis.hjust}{Numeric. The horizontal justification of the x axis identities.}

\item{y.axis.title.size}{Numeric. The size of the y axis title.}

\item{legend.text.size}{Numeric. The size of the legend text. Ignored if legend is FALSE.}

\item{legend.position}{Character. Which side to display the legend, "left", "right", "top" or "bottom". Ignored if legend is FALSE.}

\item{split.plot.title.size}{Numeric. The size of the plot titles. Ignored if split.by is NULL.}

\item{legend}{Logical. If TRUE, the legend will be shown at the bottom of the ggplot object or patchwork.}

\item{percent}{Logical. If TRUE, the proportion should be shown as a percentage of total cells for each identity.}

\item{nrow}{Numeric. The number of rows in the patchwork. Ignored if the function returns a ggplot object.}

\item{unique.group.plot}{Logical. If TRUE, the stacked proportion of cells will be gathered in a single ggplot object.}

\item{unique.split.plot}{Logical. If TRUE, the ggplot objects will be gathered in a single patchwork.}
}
\value{
A ggplot object, a list of ggplot objects, a patchwork of ggplot objects or a list of patchworks of ggplot objects.
}
\description{
This function generates a barplot of the proportion of cells from a Seurat object.
}
