% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Barplot_Cell_Proportion.R
\name{Barplot_Cell_Proportion}
\alias{Barplot_Cell_Proportion}
\title{Bar plot of the proportion of cells in each identity}
\usage{
Barplot_Cell_Proportion(
  seurat_object,
  group.by = NULL,
  split.by = NULL,
  colors = NULL,
  order.proportion = NULL,
  order.group = NULL,
  order.split = NULL,
  order.colors = TRUE,
  alpha = 1,
  show.cellsum.label = TRUE,
  cellsum.label.size = 3,
  axis.text.size = 9,
  x.axis.angle = 60,
  x.axis.hjust = 1,
  y.axis.title.size = 11,
  legend.text.size = 9,
  legend.side = "bottom",
  show.legend = TRUE,
  split.plot.title.size = 24,
  percent = TRUE,
  nrow = 1,
  unique.group.plot = TRUE,
  unique.split.plot = FALSE
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{group.by}{Character. The name of an identity in the meta.data slot to group the active.ident identity by into a stacked barplot.}

\item{split.by}{Character. The name of an identity in the meta.data slot to split the active.ident identity by into separate ggplot objects.}

\item{colors}{Character. A vector of colors to use for the active.ident identity, of same length as the number of identities in the active.ident identity. If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{order.proportion}{Character or Numeric. A vector specifying either 'reverse' or the levels (as character or as numeric values corresponding to the indexes) of the active.ident identity to order the cells.}

\item{order.group}{Character. A vector specifying either 'reverse' or the levels of the \code{group.by} identity to order the cells. Ignored if \code{group.by} = \code{NULL}.}

\item{order.split}{Character. A vector specifying either 'reverse' or the levels of the \code{split.by} identity to order the cells. Ignored if \code{split.by} = \code{NULL}.}

\item{order.colors}{Logical. If \code{TRUE}, the colors for the active.ident identity, the \code{group.by} identity and the \code{split.by} identity will automatically be ordered according to \code{order.idents}, \code{order.group} and \code{order.split}. Ignored if \code{order.idents}, \code{order.group} and \code{order.split} are \code{NULL}.}

\item{alpha}{Numeric. The transparency of the bars colors. A value between 0 and 1.}

\item{show.cellsum.label}{Logical. If \code{TRUE}, the cell sum of each identity will be shown at the top of each bar.}

\item{cellsum.label.size}{Numeric. The font size of the cell sum label. Ignored if \code{show.cellsum.label} = \code{FALSE}.}

\item{axis.text.size}{Numeric. The font size of the identities names and cell percent or numbers.}

\item{x.axis.angle}{Numeric. The angle of rotation of the identities names.}

\item{x.axis.hjust}{Numeric. The horizontal justification of the identities names.}

\item{y.axis.title.size}{Numeric. The font size of the y axis title.}

\item{legend.text.size}{Numeric. The font size of the legend text. Ignored if \code{show.legend} = \code{FALSE}.}

\item{legend.side}{Character. The side where the legend will be displayed, either 'left', 'right', 'top' or 'bottom'. Ignored if \code{show.legend} = \code{FALSE}.}

\item{show.legend}{Logical. If \code{TRUE}, shows the legend.}

\item{split.plot.title.size}{Numeric. The font size of the split plots titles. Ignored if \code{split.by} = \code{NULL}.}

\item{percent}{Logical. If \code{TRUE}, the proportion of cells will be shown as a percentage of total cells for each identity.}

\item{nrow}{Numeric. The number of rows in the patchwork. Ignored if \code{group.by} = \code{NULL} and \code{split.by} = \code{NULL}.}

\item{unique.group.plot}{Logical. If \code{TRUE}, the stacked barplots will be gathered in a single ggplot object.}

\item{unique.split.plot}{Logical. If \code{TRUE}, the ggplot objects will be gathered in a single patchwork.}
}
\value{
A ggplot object, a list of ggplot objects, a patchwork of ggplot objects or a list of patchworks of ggplot objects.
}
\description{
This function generates a barplot of the proportion of cells from a \pkg{Seurat} object.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")
\dontshow{
pbmc3k = subset(pbmc3k, idents = "Platelet", invert = TRUE)
}

# Example 1: default parameters
Barplot_Cell_Proportion(pbmc3k,
                        percent = FALSE)
}
