% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveSmoothers.R
\name{curveSmoothers}
\alias{curveSmoothers}
\title{Scatterplot of log-transformed counts, and average fitted GAM smoother curves along pseudotime}
\usage{
curveSmoothers(
  models,
  predictSmooth.df = NULL,
  genes = if (is.null(predictSmooth.df)) NA else unique(predictSmooth.df$gene),
  lineages,
  conditions = NULL,
  facets = c("genes", "conditions"),
  nPoints = 100,
  branch.points = NULL,
  points.size = 3,
  points.alpha = 0.8,
  curves.width = 1,
  curves.alpha = 1,
  facets.scales = "free",
  facets.axes = "all",
  colors = NULL,
  axis.text.size = 9,
  axis.title.size = 11,
  facets.title.size = 9,
  legend.names = NULL,
  legend.text.size = 9,
  nrow = floor(sqrt(length(genes)))
)
}
\arguments{
\item{models}{A \pkg{SingleCellExperiment} object containing the fitted GAM smoothers, computed using \code{\link[tradeSeq]{fitGAM}}, with or without \code{conditions} provided.}

\item{predictSmooth.df}{A \code{data.frame} object containing the average fitted GAM smoothers, computed using \code{\link[tradeSeq]{predictSmooth}} with \code{nPoints}, \code{lineages} and \code{tidy} = \code{TRUE}. If \code{NULL}, average fitted GAM smoothers will be internally computed.}

\item{genes}{Character. The names of one or several genes to plot the log-transformed counts and the average fitted GAM smoothers from.}

\item{lineages}{Numeric. The indices of the lineages (for example, c(1, 5), 2:4 etc) to plot the log-transformed counts, the average fitted GAM smoothers and the pseudotime values from.}

\item{conditions}{Character. The names of one or several \code{conditions} identities to select. If \code{NULL}, all identities are used, and each unique condition will be plotted, for each of the \code{lineages} provided (for example, if you have four lineages and three conditions, twelve curves will be plotted). Ignored if the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}. Please note that if the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} with \code{conditions}, it is not possible to plot global lineages (without \code{conditions}), you would need to compute a new \code{models} object using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}. This is due to a limitation in how \code{\link[tradeSeq]{predictSmooth}} returns average fitted GAM smoothers (if the \code{models} object was computed with \code{conditions}, the function will always mean the fitted GAM smoothers for each lineage and each condition independently).}

\item{facets}{Character. The names of one or two metadata among 'genes', 'lineages' and/or 'conditions' to facet the plot. If one metadata is provided, \code{\link[ggplot2]{facet_wrap}} will be used, and if two are provided, \code{\link[ggplot2]{facet_grid}} will be used, with the first metadata as rows and the second as columns. If the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}, you cannot facet by 'conditions'. Please note that if the \code{models} object was computed using \code{\link[tradeSeq]{fitGAM}} with \code{conditions}, it is not possible to plot global lineages (without \code{conditions}), you would need to compute a new \code{models} object using \code{\link[tradeSeq]{fitGAM}} without \code{conditions}. This is due to a limitation in how \code{\link[tradeSeq]{predictSmooth}} returns average fitted GAM smoothers (if the \code{models} object was computed with \code{conditions}, the function will always mean the fitted GAM smoothers for each lineage and each condition independently).}

\item{nPoints}{Numeric. (from \code{\link[tradeSeq]{predictSmooth}} documentation) The number of points used to create the grid along the smoother for each lineage.}

\item{branch.points}{Numeric, Character or List. Branching points may be shown on the fitted GAM smoother curves, to partition the lineages and help visualize differences. Either one or more pseudotime values to plot the branching points at, and/or one or several values containing 'knot' followed by a number (for example, 'knot2', 'knot4' etc), which correspond to the knots (\code{k}) input in \code{\link[tradeSeq]{fitGAM}} and which divide each lineage into segments; the function will extract the pseudotime values from each knot number and plot the branching points. Mixing any of the two options is possible (for example, c(4.3, 7, 'knot4')). You may also provide a \code{list} containing any of the two options, and named after one or several identities of a single metadata among 'genes', 'lineages' or 'conditions' (for example, list('CD4' = c(2, 'knot5')) which will only plot branching points on CD4 curves). For \code{lineages}, the \code{list} names need to be provided as 'Lineage' (with capital L) followed by the index (for example, list('Lineage1' = c(4.3, 'knot4'), 'Lineage2' = c('knot2', 7, 9)), which will only plot branching points on lineage 1 and 2 curves). Finally, you may also provide identities corresponding to several metadata at the same time by pasting them together with '_' (for example, list('Lineage1_CD4' = c(1, 'knot1', 12)), which will only plot branching points on lineage 1 and CD4 curves).}

\item{points.size}{Numeric. The size of the log-transformed count points.}

\item{points.alpha}{Numeric. The transparency of the log-transformed count points.}

\item{curves.width}{Numeric. The width of the fitted GAM smoother curves.}

\item{curves.alpha}{Numeric. The transparency of the fitted GAM smoother curves.}

\item{facets.scales}{Character. (from \code{\link[ggplot2]{facet_grid}} documentation) Are scales shared across all facets (the default, "\code{fixed}"), or do they vary across rows ("\code{free_x}"), columns ("\code{free_y}"), or both rows and columns ("\code{free}")?}

\item{facets.axes}{Character. (from \code{\link[ggplot2]{facet_grid}} documentation) Determines which axes will be drawn. When "\code{margins}" (default), axes will be drawn at the exterior margins. "\code{all_x}" and "\code{all_y}" will draw the respective axes at the interior panels too, whereas "\code{all}" will draw all axes at all panels.}

\item{colors}{Character. The color names for each identity of the metadata excluded from \code{facets} among 'genes', 'lineages' and/or 'conditions' (for example, if \code{facets} = c('genes', 'conditions'), the \code{colors} will be for each of the \code{lineages} provided, or if \code{facets} = 'lineages', the \code{colors} will be for each of the \code{genes} times \code{conditions} provided). If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{axis.text.size}{Numeric. The font size of the pseudotime and log-transformed counts.}

\item{axis.title.size}{Numeric. The font size of the axis title.}

\item{facets.title.size}{Numeric. The font size of the facet titles.}

\item{legend.names}{Character. You may provide custom names for the colors displayed in the legend, instead of the default names (for example, if \code{facets} = c('genes', 'conditions'), the color names will be 'Lineage' followed by the index of the \code{lineages} provided, you may replace these with any other names).}

\item{legend.text.size}{Numeric. The font size of the legend text.}

\item{nrow}{Numeric. The number of rows to use for the facets. Ignored if two metadata are provided to \code{facets}.}
}
\value{
A \pkg{ggplot2} object.
}
\description{
This function generates a scatterplot of the log-transformed counts, and displays on top the average fitted GAM smoother curves, computed using \code{\link[tradeSeq]{fitGAM}}, along pseudotime, and is a reworked version of \code{\link[tradeSeq]{plotSmoothers}}, it allows for the plotting of multiple genes simultaneously, as well as selected lineages and conditions. Additionally, the scatterplot is rasterized to reduce the size of the output file and branching points may be displayed on each curve.
}
