% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid_VlnPlot.R
\name{Grid_VlnPlot}
\alias{Grid_VlnPlot}
\title{Square stacked violin plot of gene expression in each identity}
\usage{
Grid_VlnPlot(
  seurat_object,
  assay = "RNA",
  layer = "data",
  features,
  idents = NULL,
  scale = TRUE,
  rotate.axis = FALSE,
  colors = NULL,
  order.idents = NULL,
  order.colors = TRUE,
  idents.text.size = 9,
  show.idents = FALSE,
  features.text.size = 9,
  legend.text.size = 7,
  legend.side = "bottom",
  show.legend = TRUE,
  ncol = "square"
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{assay}{Character. If the \pkg{Seurat} object contains multiple RNA assays, you may specify which one to use (for example 'RNA2' if you have created a second RNA assay you named 'RNA2'. See \href{https://satijalab.org/seurat/articles/seurat5_essential_commands.html#create-seurat-or-assay-objects}{Seurat v5 vignettes} for more information). You may also use another assay such as 'SCT' to pull features expression from.}

\item{layer}{Character. Formerly known as slot. It is recommended to use 'data'.}

\item{features}{Character. A vector of features to plot.}

\item{idents}{Character. A vector with one or several identities names in the active.ident identity to use if you only want those (instead of subsetting your object). If \code{NULL}, all identities will be used.}

\item{scale}{Logical. If \code{TRUE}, scales the violins to have the same max height between features.}

\item{rotate.axis}{Logical. If \code{TRUE}, flips the axis, displaying violins vertically instead of horizontally.}

\item{colors}{Character. A vector of colors to use for the active.ident identity, of same length as the number of identities in the active.ident identity or supplied to the \code{idents} parameter. If \code{NULL}, uses \pkg{Seurat}'s default colors.}

\item{order.idents}{Character or Numeric. A vector specifying either 'reverse' or the levels (as character or as numeric values corresponding to the indexes) of the active.ident identity to order the cells.}

\item{order.colors}{Logical. If \code{TRUE}, the colors for the active.ident identity will automatically be ordered according to \code{order.idents}. Ignored if \code{order.idents} = \code{NULL}.}

\item{idents.text.size}{Numeric. The font size of the identities names. Ignored if \code{show.idents} = \code{FALSE}.}

\item{show.idents}{Logical. If \code{TRUE}, shows the identities names on the plot.}

\item{features.text.size}{Numeric. The font size of the features names.}

\item{legend.text.size}{Numeric. The font size of the legend text. Ignored if \code{show.legend} = \code{FALSE}.}

\item{legend.side}{Character. The side where the legend will be displayed, either 'left', 'right', 'top' or 'bottom'. Ignored if \code{show.legend} = \code{FALSE}.}

\item{show.legend}{Logical. If \code{TRUE}, shows the legend.}

\item{ncol}{Numeric. Number of columns to use. If 'square', will display features in a square grid or as close as possible depending on number of features.}
}
\value{
A ggplot object.
}
\description{
This function is a stacked violin plot optimized to display features expression in a \pkg{Seurat} object in a grid fashion (square) instead of a single column like other stacked violin functions available in other packages, resulting in nicer plots and easier to include in publications.
}
\examples{
\dontshow{
suppressWarnings(suppressPackageStartupMessages(library(Seurat)))
}
# Prepare data
pbmc3k <- Right_Data("pbmc3k")
pbmc3k.markers <- c("CCR7", "CD14", "CD40LG",
                    "CD79A", "CD8A", "CDKN1C",
                    "GNLY", "CLEC10A", "PPBP")

# Example 1: default parameters
Grid_VlnPlot(pbmc3k,
             features = pbmc3k.markers)
}
