% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Right_DietSeurat.R
\name{Right_DietSeurat}
\alias{Right_DietSeurat}
\title{Reduce the size of a Seurat object}
\usage{
Right_DietSeurat(
  seurat_object,
  idents = NULL,
  cells = NULL,
  features = NULL,
  dimreducs = NULL,
  graphs = NULL,
  variable.features = FALSE,
  misc = TRUE,
  split.counts.layer = NULL,
  data.layer = FALSE,
  scale.layer = FALSE,
  SCTAssay = FALSE
)
}
\arguments{
\item{seurat_object}{A \pkg{Seurat} object.}

\item{idents}{Character. The names of one or several metadata (for example, 'orig.ident', 'seurat_clusters', etc) to keep in the diet object. If \code{NULL}, all metadata will be kept.}

\item{cells}{Character. The names of one or several cells to keep in the diet object. If \code{NULL}, all cells will be kept.}

\item{features}{Character. The names of one or several features to keep in the diet object. If \code{NULL}, all features will be kept.}

\item{dimreducs}{Character. The names of one or several reductions to keep in the diet object. If \code{NULL}, all reductions will be removed. Please note that if you subset features and remove all from which principal components (PC) were calculated, the 'PCA' reduction will not be kept, even if it is provided, as the feature.loadings slot will be empty. This is intended behavior.}

\item{graphs}{Character. The names of one or several graphs to keep in the diet object. If \code{NULL}, all graphs will be removed.}

\item{variable.features}{Logical. If \code{TRUE}, the variable features slot will be kept in the diet object. Please note that if you keep only a subset of features, the variable features will also be subset, and if you remove all features from which variable features were calculated, the variable features will not be kept even if \code{TRUE}. This is intended behavior.}

\item{misc}{Logical. If \code{TRUE}, the misc slot will be kept in the diet object.}

\item{split.counts.layer}{Character. The name of a metadata (for example, 'orig.ident', 'seurat_clusters', etc) to split the 'counts' layer in the 'RNA' assay by if you need to in your downstream analysis. If \code{NULL}, the diet object will have a single 'counts' layer, even if the original object had split 'counts' layers.}

\item{data.layer}{Logical. If \code{TRUE}, the 'data' layer in the 'RNA' assay will be kept in the diet object if it is present. As with the 'counts' layer, if there are split 'data' layers, they will be joined into a single 'data' layer unless \code{split.counts.layer} is provided.}

\item{scale.layer}{Logical. If \code{TRUE}, the 'scale.data' layer in the 'RNA' assay will be kept in the diet object if it is present.}

\item{SCTAssay}{Logical. If \code{TRUE}, the 'SCT' assay will be kept in the diet object if it is present.}
}
\value{
A \pkg{Seurat} object, hopefully smaller, with class Assay5 'RNA' assay and specified layers and slots.
}
\description{
As of \pkg{Seurat} v5 release, \code{\link[Seurat]{DietSeurat}} does not remove data and scale.data layers, resulting in objects with little to no reduction in size. This function was created with the purpose to restore \code{\link[Seurat]{DietSeurat}}'s proper behavior until the function is fixed by \pkg{Seurat}'s dev team, as well as to offer a few new options such as the ability to subset cells or keep variable features.
}
